# docs/swagger.yaml
definitions:
  model.CVTemplate:
    properties:
      category:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      deleted_by:
        type: integer
      description:
        type: string
      download_count:
        type: integer
      file_path:
        type: string
      file_type:
        description: docx, xlsx
        type: string
      id:
        type: integer
      is_deleted:
        type: boolean
      name:
        type: string
      tags:
        type: string
      thumbnail_path:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
  response.VResponse:
    properties:
      code: {}
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /datn_backend/cv/admin/templates:
    post:
      consumes:
      - multipart/form-data
      description: Tải lên file mẫu CV mới
      parameters:
      - description: CV Template file (.docx, .xlsx)
        in: formData
        name: template_file
        required: true
        type: file
      - description: Template name
        in: formData
        name: name
        required: true
        type: string
      - description: Template description
        in: formData
        name: description
        type: string
      - description: Template category
        in: formData
        name: category
        type: string
      - description: Template tags
        in: formData
        name: tags
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.VResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CVTemplate'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.VResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Tạo mẫu CV mới
      tags:
      - Admin
  /datn_backend/cv/admin/templates/{id}:
    delete:
      consumes:
      - application/json
      description: Xóa mẫu CV
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.VResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.VResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Xóa mẫu CV
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin mẫu CV
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      - description: Template info
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/model.CVTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.VResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CVTemplate'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.VResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.VResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật mẫu CV
      tags:
      - Admin
  /datn_backend/cv/admin/templates/{id}/download-original:
    get:
      consumes:
      - application/json
      description: Tải xuống file mẫu CV gốc (không điền thông tin)
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: CV Template file
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Tải xuống mẫu CV gốc
      tags:
      - Admin
  /datn_backend/cv/admin/templates/{id}/thumbnail:
    post:
      consumes:
      - multipart/form-data
      description: Tải lên hoặc cập nhật thumbnail cho mẫu CV
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      - description: Thumbnail file (.jpg, .jpeg, .png, .gif)
        in: formData
        name: thumbnail_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.VResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CVTemplate'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.VResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.VResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Tải lên thumbnail cho mẫu CV
      tags:
      - Admin
  /datn_backend/cv/templates:
    get:
      consumes:
      - application/json
      description: Lấy danh sách tất cả các mẫu CV
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.VResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.CVTemplate'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      summary: Lấy tất cả mẫu CV
      tags:
      - CV
  /datn_backend/cv/templates/{id}:
    get:
      consumes:
      - application/json
      description: Lấy thông tin chi tiết của mẫu CV theo ID
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.VResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.CVTemplate'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      summary: Lấy mẫu CV theo ID
      tags:
      - CV
  /datn_backend/cv/templates/{id}/download:
    get:
      consumes:
      - application/json
      description: Tải xuống file mẫu CV với thông tin cá nhân được điền vào
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: CV Template file with personal info
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.VResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      security:
      - BearerAuth: []
      summary: Tải xuống mẫu CV với thông tin cá nhân
      tags:
      - CV
  /datn_backend/cv/templates/{id}/preview:
    get:
      consumes:
      - application/json
      description: Xem trước hình ảnh thumbnail của mẫu CV
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/jpeg
      responses:
        "200":
          description: CV Template thumbnail
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.VResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.VResponse'
      summary: Xem trước mẫu CV
      tags:
      - CV
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"